+++ src/Service/Yaml/KeysParser.php
@@ -21,7 +21,6 @@ final class KeysParser
     public function getOmittedKeys(array $localesKeys): array
     {
         $allDomainKeys = $this->mergeKeys($localesKeys);
-        ksort($localesKeys);
         $omittedKeys = [];
         foreach ($localesKeys as $locale => $localeKeys) {
             $omittedKeys[$locale] = array_values(array_diff($allDomainKeys, $localeKeys));
@@ -39,7 +38,7 @@ final class KeysParser
     {
         $keys = [];
         foreach ($localesFiles as $locale => $files) {
-            $keys[$locale] = $this->getSortedKeys($this->parseFiles($files));
+            $keys[$locale] = array_keys($this->fileParser->parseFiles($files));
         }
 
         return $keys;
@@ -59,21 +58,14 @@ final class KeysParser
     }
 
     /**
+     * @deprecated since version 2.7.0. Use {@see \Aeliot\Bundle\TransMaintain\Service\Yaml\FileToSingleLevelArrayParser::parseFiles() }
+     *
      * @param string[] $files
      *
      * @return array<string,string>
      */
     public function parseFiles(array $files): array
     {
-        // THINK: sort result
-        return array_merge(...array_map(fn (string $x): array => $this->fileParser->parse($x), array_values($files)));
-    }
-
-    private function getSortedKeys(array $values): array
-    {
-        $keys = array_keys($values);
-        sort($keys);
-
-        return $keys;
+        return $this->fileParser->parseFiles($files);
     }
 }