+++ tests/Unit/Model/ReportBagTest.php
@@ -59,7 +59,7 @@ final class ReportBagTest extends TestCase
 
     public function testLineColumnNames(): void
     {
-        //TODO implement testing if report line columns is same to configured
+        // TODO implement testing if report line columns is same to configured
         $this->markTestSkipped('Test not implemented yet');
     }