+++ tests/Unit/Service/Yaml/Transformer/KeysSorterTest.php
@@ -5,7 +5,6 @@ declare(strict_types=1);
 namespace Aeliot\Bundle\TransMaintain\Test\Unit\Service\Yaml\Transformer;
 
 use Aeliot\Bundle\TransMaintain\Service\Yaml\Transformer\KeysSorter;
-use Generator;
 use PHPUnit\Framework\TestCase;
 
 final class KeysSorterTest extends TestCase
@@ -18,7 +17,7 @@ final class KeysSorterTest extends TestCase
         self::assertSame($expected, (new KeysSorter())->transform($income));
     }
 
-    public function getDataForTestTransform(): Generator
+    public function getDataForTestTransform(): \Generator
     {
         yield [['a' => '*', 'b' => '*'], ['a' => '*', 'b' => '*']];
         yield [['a' => '*', 'b' => '*'], ['b' => '*', 'a' => '*']];